SUB-ENTITY/SUPER-ENTITY EXPERIMENT BEGINS HERE
[#166580361] https://coolambience.com site

[Adhoc]
 \
  * It's XJ Music (TM) until we get the certificate

[#167044451] Test suite should execute entire Java test suite without stopping
 \
  * [core/CoreIT] Integration tests inherit @After method to call shutdown hook after test

EDIT THESE NOTES All tests of Java modules pass, however:

[#166598919] Edit Chain, Program + Instrument in v2 UX
 \
  * [ui/hub-ui] Stub Program and Instrument sub-entity editing for now, pending implementation of program-ui and instrument-ui

[#167585908] Correct fabrication with v2 core
 \
  * [core/ingest/cache] invalidation should not halt worker

[#150279494] On mobile, open menu and make selection, expect menu to close

[#167585930] UI shows errors returned in payload

[#166708597] Instrument model handles all of its own entities
  + Audio model handles all of its own entities, ala Segments #166132897
  + Audio Child Entities are identified and related by UUID (not id), ala Segments #166273140
  + Audio is an interface provided by a PatternFactory, ala Segment #166317849
  + Audio have no child SQL entities, only content JSON, ala Segment #166109389
  + Audio Child entities have UUID (not id) appropriately non-dependent on SQL-like entity id, ala Segment #166273140
  + MIGRATION TRANSFORMS PRE-EXISTING AUDIOS INTO NEW JSON OBJECTS!
  + Deprecate AudioErase job
  + Deprecate AudioClone job
  + Audio erasure on S3 will be moved to a new universal garbage collection job
  + Set all content of instrument, cloned from another source instrument, with all new UUID, preserving relationships.
  + Instrument doesn't actually have density property; it's computed by averaging the density of all its sub entities
  + JSON Serialization + Deserialization unit tests for meme, audio, audio_chord, and audio_event

[#166690830] Program model handles all of its own entities
  + Program model handles all of its own entities, ala Segments #166132897
  + Program Child Entities are identified and related by UUID (not id), ala Segments #166273140
  + Program is an interface provided by a PatternFactory, ala Segment #166317849
  + Program have no child SQL entities, only content JSON, ala Segment #166109389
  + Child entities have UUID (not id) appropriately non-dependent on SQL-like entity id, ala Segment #166273140
  + MIGRATION TRANSFORMS PRE-EXISTING AUDIOS INTO NEW JSON OBJECTS!
  + Deprecate PatternClone job
  + Deprecate PatternErase job
  + Set all content of program, cloned from another source program, with all new UUID, preserving relationships.
  + Pattern has voice (Deprecates: pattern event having voice)
  + replace fetch one Pattern with usage of Program internal content
  + Choice.sequenceId() sequence_id migrated to Choice.programId()
  + Deprecate Choice validation requirement "Cannot have both Sequence ID and Sequence Pattern ID."
  + Deprecate Choice validation requirement "Required to have either Sequence ID or Sequence Pattern ID."
  + Choice validation always "Required to have a Program ID."
  + figure out this confusion, rhythm programs are now going to need to consider sequence bindings?
  + Rhythm and Detail programs are allowed to have only one (default) sequence.
  + Sequence has total now!
  + Program doesn't actually have density property; it's computed by averaging the density of all its sub entities
  + JSON Serialization + Deserialization unit tests for meme, pattern, pattern_event, sequence, sequence_binding, sequence_binding_meme, sequence_chord, and voice

[#167282842] Chain handles all of its own binding + config entities

[#167276586] JSON API facilitates complex transactions
 \
  * [core/model/payload]
      + PayloadDeserializer uses Jackson; we're locked into that because that's what Jersey (our REST framework) uses.
      + PayloadSerializer uses GSON for outbound data, because it's simpler.

[Adhoc]
 \
  * [ui/hub-ui] Upgrade to Ember 3.11
  * [core] require migration of legacy tables to new child entities
  * [hub] entirely missing endpoints for Program, Instrument, ChainBinding, and ChainConfig
  * [ui/hub] missing support for new Program, Instrument, ChainBinding, and ChainConfig
  * [core/persistence/sql] SQLDatabaseProvider implements javax.DataSource (deprecating DriverManager)

+++ core/src/main/java/io/xj/core/util/TremendouslyRandom.java
@@ -27,7 +27,7 @@ public interface TremendouslyRandom {
    */
   static Double zeroToLimit(Double limit) {
     if ((double) 0 >= limit) return 0.0;
-    return ThreadLocalRandom.current().nextDouble((double) 0, limit);
+    return ThreadLocalRandom.current().nextDouble(0, limit);
   }
 
 }
